refactor: Move file and readline includes to common directory and remove unused headers
