+++ include/openshieldhit.h
@@ -18,12 +18,10 @@
 #include "openshieldhit/beam.h"
 #include "openshieldhit/beam_defs.h"
 #include "openshieldhit/const.h"
-#include "openshieldhit/file.h"
 #include "openshieldhit/geometry.h"
 #include "openshieldhit/geometry_defs.h"
 #include "openshieldhit/logger.h"
 #include "openshieldhit/material.h"
-#include "openshieldhit/readline.h"
 #include "openshieldhit/scoring.h"
 #include "openshieldhit/simulation.h"
 #include "openshieldhit/status.h"
