+++ src/apps/osh/osh_app_osh.c
@@ -8,7 +8,7 @@
 #include "apps/osh/osh_geometry_parse.h"
 #include "apps/osh/osh_material_parse.h"
 #include "apps/osh/osh_scoring_parse.h"
-#include "openshieldhit/file.h"
+#include "common/osh_file.h"
 #include "openshieldhit/logger.h"
 #include "openshieldhit/material.h"
 #include "openshieldhit/status.h"
