+++ src/apps/osh/osh_material_parse.h
@@ -1,7 +1,7 @@
 #ifndef OSH_FRONTEND_OPENSHIELDHIT_MATERIAL_PARSE_H
 #define OSH_FRONTEND_OPENSHIELDHIT_MATERIAL_PARSE_H
 
-#include "openshieldhit/file.h"
+#include "common/osh_file.h"
 #include "openshieldhit/material.h"
 #include "openshieldhit/status.h"
