+++ src/apps/osh/osh_scoring_parse_output.c
@@ -19,8 +19,8 @@
 
 #include "apps/osh/osh_scoring_parse_internal.h"
 #include "apps/osh/osh_scoring_parse_keys.h"
+#include "common/osh_readline.h"
 #include "openshieldhit/logger.h"
-#include "openshieldhit/readline.h"
 
 typedef enum osh_status (*output_handler_fn)(struct osh_scoring_output_def *, char **, int, char const *, unsigned int);
